us uint as ID

+++ Backend/api/user.go
@@ -5,10 +5,10 @@ import (
 )
 
 type UserRepository interface {
-	CreateUser(username, email, passwordHash, walletAddress string) (int64, error)
-	GetUserByID(id int64) (*models.User, error)
+	CreateUser(username, email, passwordHash, walletAddress string) (uint, error)
+	GetUserByID(id uint) (*models.User, error)
 	GetUserByUsername(username string) (*models.User, error)
 	GetUserByWalletAddress(walletAddress string) (*models.User, error)
 	UpdateUser(id uint, updates map[string]interface{}) error
-	DeleteUser(id int64) error
+	DeleteUser(id uint) error
 }
